split all over files in a node folder
