+++ packages/mobx-state-tree/src/core/node/identifier-cache.ts
@@ -1,7 +1,7 @@
-import { fail } from "../utils"
+import { fail } from "../../utils"
 import { observable, IObservableArray } from "mobx"
-import { IType } from "../types/type"
-import { INode } from "./node"
+import { IType } from "../../types/type"
+import { INode } from "./node-utils"
 
 export class IdentifierCache {
     private cache = observable.map<IObservableArray<INode>>()
