+++ packages/mobx-state-tree/src/core/node/scalar-node.ts
@@ -1,73 +1,5 @@
-import { observable, computed, reaction } from "mobx"
-
 let nextNodeId = 1
-export interface INode {
-    readonly type: IType<any, any>
-
-    readonly storedValue: any
-    readonly _environment: any
-    readonly path: string
-    readonly isRoot: boolean
-    readonly parent: INode | null
-    readonly root: INode
-    subpath: string
-
-    isRunningAction(): boolean
-    _isRunningAction: boolean
-    isProtectionEnabled: boolean
-    isProtected: boolean
-    assertWritable(): void
-
-    identifier: string | null
-
-    isAlive: boolean
-    assertAlive(): void
-
-    identifierAttribute: string | undefined
-    identifierCache: IdentifierCache | undefined
-
-    readonly value: any
-    readonly snapshot: any
-
-    resolve(pathParts: string): INode
-    resolve(pathParts: string, failIfResolveFails: boolean): INode | undefined
-    resolve(path: string, failIfResolveFails?: boolean): INode | undefined
-
-    resolvePath(pathParts: string[]): INode
-    resolvePath(pathParts: string[], failIfResolveFails: boolean): INode | undefined
-    resolvePath(pathParts: string[], failIfResolveFails?: boolean): INode | undefined
-
-    getRelativePathTo(target: INode): string
-    setParent(newParent: INode | null, subpath: string | null): void
-
-    getChildNode(subpath: string): INode
-    getChildren(): INode[]
-    getChildType(key: string): IType<any, any>
-    removeChild(subpath: string): void
-
-    unbox(childNode: INode): any
-    detach(): void
-
-    die(): void
-    aboutToDie(): void
-    finalizeDeath(): void
-
-    emitPatch(basePatch: IReversibleJsonPatch, source: INode): void
-
-    readonly middlewares: IMiddlewareHandler[]
-
-    addMiddleWare(handler: IMiddlewareHandler): IDisposer
-
-    applyPatchLocally(subpath: string, patch: IJsonPatch): void
-    onPatch(handler: (patch: IJsonPatch, reversePatch: IJsonPatch) => void): IDisposer
-    applyPatches(patches: IJsonPatch[]): void
-
-    applySnapshot(snapshot: any): void
-    onSnapshot(onChange: (snapshot: any) => void): IDisposer
-    addDisposer(disposer: () => void): void
-}
-
-export class ImmutableNode implements INode {
+export class ScalarNode implements INode {
     // optimization: these fields make MST memory expensive for primitives. Most can be initialized lazily, or with EMPTY_ARRAY on prototype
     readonly nodeId = ++nextNodeId
     readonly type: IType<any, any>
@@ -409,273 +341,24 @@ export class ImmutableNode implements INode {
     }
 }
 
-export class Node extends ImmutableNode implements INode {
-    middlewares: IMiddlewareHandler[]
-    private snapshotSubscribers: ((snapshot: any) => void)[]
-    // TODO: split patches in two; patch and reversePatch
-    private patchSubscribers: ((patch: IJsonPatch, reversePatch: IJsonPatch) => void)[]
-    private disposers: (() => void)[]
-    // applyPatches: (patches: IJsonPatch[]) => void
-    // applySnapshot: (snapshot: any) => void
-
-    constructor(
-        type: IType<any, any>,
-        parent: INode | null,
-        subpath: string,
-        environment: any,
-        initialValue: any,
-        storedValue: any,
-        canAttachTreeNode: boolean,
-        finalizeNewInstance: (node: INode, initialValue: any) => void = noop
-    ) {
-        super(
-            type,
-            parent,
-            subpath,
-            environment,
-            initialValue,
-            storedValue,
-            canAttachTreeNode,
-            finalizeNewInstance
-        )
-
-        // optimization: don't keep the snapshot by default alive with a reaction by default
-        // in prod mode. This saves lot of GC overhead (important for e.g. React Native)
-        // if the feature is not actively used
-        // downside; no structural sharing if getSnapshot is called incidently
-        const snapshotDisposer = reaction(
-            () => this.snapshot,
-            snapshot => {
-                this.emitSnapshot(snapshot)
-            }
-        )
-        snapshotDisposer.onError((e: any) => {
-            throw e
-        })
-        this.addDisposer(snapshotDisposer)
-    }
-
-    preboot() {
-        this.disposers = []
-        this.middlewares = []
-        this.snapshotSubscribers = []
-        this.patchSubscribers = []
-
-        // Optimization: this does not need to be done per instance
-        // if some pieces from createActionInvoker are extracted
-        this.applyPatches = createActionInvoker(
-            this.storedValue,
-            "@APPLY_PATCHES",
-            (patches: IJsonPatch[]) => {
-                patches.forEach(patch => {
-                    const parts = splitJsonPath(patch.path)
-                    const node = this.resolvePath(parts.slice(0, -1))
-                    node.applyPatchLocally(parts[parts.length - 1], patch)
-                })
-            }
-        ).bind(this.storedValue)
-        this.applySnapshot = createActionInvoker(
-            this.storedValue,
-            "@APPLY_SNAPSHOT",
-            (snapshot: any) => {
-                // if the snapshot is the same as the current one, avoid performing a reconcile
-                if (snapshot === this.snapshot) return
-                // else, apply it by calling the type logic
-                return this.type.applySnapshot(this, snapshot)
-            }
-        ).bind(this.storedValue)
-    }
-
-    public die() {
-        if (this._isDetaching) return
-
-        if (isStateTreeNode(this.storedValue)) {
-            walk(this.storedValue, child => getStateTreeNode(child).aboutToDie())
-            walk(this.storedValue, child => getStateTreeNode(child).finalizeDeath())
-        }
-    }
-
-    public aboutToDie() {
-        this.disposers.splice(0).forEach(f => f())
-        this.fireHook("beforeDestroy")
-    }
-
-    public finalizeDeath() {
-        // invariant: not called directly but from "die"
-        this.root.identifierCache!.notifyDied(this)
-        const self = this
-        const oldPath = this.path
-        addReadOnlyProp(this, "snapshot", this.snapshot) // kill the computed prop and just store the last snapshot
-
-        this.patchSubscribers.splice(0)
-        this.snapshotSubscribers.splice(0)
-        this.patchSubscribers.splice(0)
-        this._isAlive = false
-        this._parent = null
-        this.subpath = ""
-
-        // This is quite a hack, once interceptable objects / arrays / maps are extracted from mobx,
-        // we could express this in a much nicer way
-        // TODO: should be possible to obtain id's still...
-        Object.defineProperty(this.storedValue, "$mobx", {
-            get() {
-                fail(
-                    `This object has died and is no longer part of a state tree. It cannot be used anymore. The object (of type '${self
-                        .type
-                        .name}') used to live at '${oldPath}'. It is possible to access the last snapshot of this object using 'getSnapshot', or to create a fresh copy using 'clone'. If you want to remove an object from the tree without killing it, use 'detach' instead.`
-                )
-            }
-        })
-    }
-
-    public onSnapshot(onChange: (snapshot: any) => void): IDisposer {
-        return registerEventHandler(this.snapshotSubscribers, onChange)
-    }
-
-    public emitSnapshot(snapshot: any) {
-        this.snapshotSubscribers.forEach((f: Function) => f(snapshot))
-    }
-
-    public onPatch(handler: (patch: IJsonPatch, reversePatch: IJsonPatch) => void): IDisposer {
-        return registerEventHandler(this.patchSubscribers, handler)
-    }
-
-    emitPatch(basePatch: IReversibleJsonPatch, source: INode) {
-        if (this.patchSubscribers.length) {
-            const localizedPatch: IReversibleJsonPatch = extend({}, basePatch, {
-                path: source.path.substr(this.path.length) + "/" + basePatch.path // calculate the relative path of the patch
-            })
-            const [patch, reversePatch] = splitPatch(localizedPatch)
-            this.patchSubscribers.forEach(f => f(patch, reversePatch))
-        }
-        if (this.parent) this.parent.emitPatch(basePatch, source)
-    }
-
-    addDisposer(disposer: () => void) {
-        this.disposers.unshift(disposer)
-    }
-
-    addMiddleWare(handler: IMiddlewareHandler) {
-        return registerEventHandler(this.middlewares, handler)
-    }
-
-    applyPatchLocally(subpath: string, patch: IJsonPatch): void {
-        this.assertWritable()
-        this.type.applyPatchLocally(this, subpath, patch)
-    }
-}
-
-export function isNode(value: any): value is INode {
-    return value instanceof Node || value instanceof ImmutableNode
-}
-
-export type IStateTreeNode = {
-    readonly $treenode?: any
-}
-
-/**
- * Returns true if the given value is a node in a state tree.
- * More precisely, that is, if the value is an instance of a
- * `types.model`, `types.array` or `types.map`.
- *
- * @export
- * @param {*} value
- * @returns {value is IStateTreeNode}
- */
-export function isStateTreeNode(value: any): value is IStateTreeNode {
-    return !!(value && value.$treenode)
-}
-
-export function getStateTreeNode(value: IStateTreeNode): INode {
-    if (isStateTreeNode(value)) return value.$treenode!
-    else return fail(`Value ${value} is no MST Node`)
-}
-
-function canAttachNode(value: any) {
-    return (
-        value &&
-        typeof value === "object" &&
-        !(value instanceof Date) &&
-        !isStateTreeNode(value) &&
-        !Object.isFrozen(value)
-    )
-}
-
-function toJSON(this: IStateTreeNode) {
-    return getStateTreeNode(this).snapshot
-}
-
-export function createNode<S, T>(
-    type: IType<S, T>,
-    parent: INode | null,
-    subpath: string,
-    environment: any,
-    initialValue: any,
-    createNewInstance: (initialValue: any) => T = identity,
-    finalizeNewInstance: (node: INode, initialValue: any) => void = noop
-) {
-    if (isStateTreeNode(initialValue)) {
-        const targetNode = getStateTreeNode(initialValue)
-        if (!targetNode.isRoot)
-            fail(
-                `Cannot add an object to a state tree if it is already part of the same or another state tree. Tried to assign an object to '${parent
-                    ? parent.path
-                    : ""}/${subpath}', but it lives already at '${targetNode.path}'`
-            )
-        targetNode.setParent(parent, subpath)
-        return targetNode
-    }
-
-    const storedValue = createNewInstance(initialValue)
-    const canAttachTreeNode = canAttachNode(storedValue)
-
-    if (canAttachTreeNode) {
-        // tslint:disable-next-line:no_unused-variable
-        return new Node(
-            type,
-            parent,
-            subpath,
-            environment,
-            initialValue,
-            storedValue,
-            canAttachTreeNode,
-            finalizeNewInstance
-        )
-    }
-    return new ImmutableNode(
-        type,
-        parent,
-        subpath,
-        environment,
-        initialValue,
-        storedValue,
-        canAttachTreeNode,
-        finalizeNewInstance
-    )
-}
-
-import { IType } from "../types/type"
+import { observable, action, computed } from "mobx"
+import { INode, toJSON } from "./node-utils"
+import { IdentifierCache } from "./identifier-cache"
 import {
     escapeJsonPath,
     splitJsonPath,
-    joinJsonPath,
-    IReversibleJsonPatch,
-    stripPatch,
     IJsonPatch,
-    splitPatch
-} from "./json-patch"
-import { walk } from "./mst-operations"
-import { IMiddlewareHandler, createActionInvoker } from "./action"
+    IReversibleJsonPatch,
+    joinJsonPath
+} from "../json-patch"
 import {
-    addReadOnlyProp,
-    addHiddenFinalProp,
-    extend,
-    fail,
-    IDisposer,
-    registerEventHandler,
-    identity,
+    EMPTY_ARRAY,
     noop,
+    identity,
+    addHiddenFinalProp,
+    fail,
     freeze,
-    EMPTY_ARRAY
-} from "../utils"
-import { IdentifierCache } from "./identifier-cache"
+    IDisposer
+} from "../../utils"
+import { IType } from "../../types/type"
+import { IMiddlewareHandler } from "../action"
